backend/requirements.txt
# Python requirements
#
# Use from command line: pip install -r requirements.txt -t lib
#

Flask==1.0.2
google-cloud-firestore==0.31.0
